Moved: domain strip function to lib.rs
